tech worker:remove "build advanced" from menu
merged the options from the "build advanced" menu to the "build" menu,
as it is with the other 6 factions. The icon for the build advanced menu
is a castle, which is confusing, annoying, and non-intuitive.
Additionally, there are only 2 items on that menu, so I merged them into
the worker's plain build menu, as is the case for the other 6 factions.